Cast 0 to uint32_t in SetStakeEntropyBit conditional
This eliminates the compiler warning about enumerated vs
non-enumerated type in conditional.

+++ src/main.h
@@ -630,7 +630,7 @@ public:
     {
         if (nEntropyBit > 1)
             return false;
-        nFlags |= (nEntropyBit? BLOCK_STAKE_ENTROPY : 0);
+        nFlags |= (nEntropyBit ? BLOCK_STAKE_ENTROPY : (uint32_t) 0);
         return true;
     }
